requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
Flask==2.3.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
paho-mqtt==2.1.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
waitress==2.1.2
Werkzeug==2.3.6
